envelope: Move ensureCommaDelimitedAddresses
This function doesn't belong in envelope, and it is confusing that
its unit test is in an unrelated file.
