+++ header_test.go
@@ -271,65 +271,3 @@ func TestReadHeader(t *testing.T) {
 		})
 	}
 }
-
-func TestCommaDelimitedAddressLists(t *testing.T) {
-	testData := []struct {
-		have string
-		want string
-	}{
-		{
-			have: `"Joe @ Company" <joe@company.com> <other@company.com>`,
-			want: `"Joe @ Company" <joe@company.com>, <other@company.com>`,
-		},
-		{
-			have: `Joe Company <joe@company.com> <other@company.com>`,
-			want: `Joe Company <joe@company.com>, <other@company.com>`,
-		},
-		{
-			have: `Joe Company:Joey <joe@company.com> John <other@company.com>;`,
-			want: `Joe Company:Joey <joe@company.com>, John <other@company.com>;`,
-		},
-		{
-			have: `Joe Company:Joey <joe@company.com> John <other@company.com>; Jimmy John <jimmy.john@company.com>`,
-			want: `Joe Company:Joey <joe@company.com>, John <other@company.com>;`,
-		},
-		{
-			have: `Joe Company <joe@company.com> John Company <other@company.com>`,
-			want: `Joe Company <joe@company.com>, John Company <other@company.com>`,
-		},
-		{
-			have: `Joe Company <joe@company.com>,John Company <other@company.com>`,
-			want: `Joe Company <joe@company.com>,John Company <other@company.com>`,
-		},
-		{
-			have: `joe@company.com other@company.com`,
-			want: `joe@company.com, other@company.com`,
-		},
-		{
-			have: `Jimmy John <jimmy.john@company.com> joe@company.com other@company.com`,
-			want: `Jimmy John <jimmy.john@company.com>, joe@company.com, other@company.com`,
-		},
-		{
-			have: `Jimmy John <jimmy.john@company.com> joe@company.com John Company <other@company.com>`,
-			want: `Jimmy John <jimmy.john@company.com>, joe@company.com, John Company <other@company.com>`,
-		},
-		{
-			have: `<boss@nil.test> "Giant; \"Big\" Box" <sysservices@example.net>`,
-			want: `<boss@nil.test>, "Giant; \"Big\" Box" <sysservices@example.net>`,
-		},
-		{
-			have: `A Group:Ed Jones <c@a.test>,joe@where.test,John <jdoe@one.test>;`,
-			want: `A Group:Ed Jones <c@a.test>,joe@where.test,John <jdoe@one.test>;`,
-		},
-		{
-			have: `A Group:Ed Jones <c@a.test> joe@where.test John <jdoe@one.test>;`,
-			want: `A Group:Ed Jones <c@a.test>, joe@where.test, John <jdoe@one.test>;`,
-		},
-	}
-	for i := range testData {
-		v := ensureCommaDelimitedAddresses(testData[i].have)
-		if testData[i].want != v {
-			t.Fatalf("Expected %s, but got %s", testData[i].want, v)
-		}
-	}
-}
