+++ internal/stringutil/addr.go
@@ -3,6 +3,7 @@ package stringutil
 import (
 	"bytes"
 	"net/mail"
+	"strings"
 )
 
 // JoinAddress formats a slice of Address structs such that they can be used in a To or Cc header.
@@ -18,4 +19,64 @@ func JoinAddress(addrs []mail.Address) string {
 		_, _ = buf.WriteString(a.String())
 	}
 	return buf.String()
+}
+
+// EnsureCommaDelimitedAddresses is used by AddressList to ensure that address lists are properly
+// delimited.
+func EnsureCommaDelimitedAddresses(s string) string {
+	// This normalizes the whitespace, but may interfere with CFWS (comments with folding whitespace)
+	// RFC-5322 3.4.0:
+	//      because some legacy implementations interpret the comment,
+	//      comments generally SHOULD NOT be used in address fields
+	//      to avoid confusing such implementations.
+	s = strings.Join(strings.Fields(s), " ")
+
+	inQuotes := false
+	inDomain := false
+	escapeSequence := false
+	sb := strings.Builder{}
+	for _, r := range s {
+		if escapeSequence {
+			escapeSequence = false
+			sb.WriteRune(r)
+			continue
+		}
+		if r == '"' {
+			inQuotes = !inQuotes
+			sb.WriteRune(r)
+			continue
+		}
+		if inQuotes {
+			if r == '\\' {
+				escapeSequence = true
+				sb.WriteRune(r)
+				continue
+			}
+		} else {
+			if r == '@' {
+				inDomain = true
+				sb.WriteRune(r)
+				continue
+			}
+			if inDomain {
+				if r == ';' {
+					sb.WriteRune(r)
+					break
+				}
+				if r == ',' {
+					inDomain = false
+					sb.WriteRune(r)
+					continue
+				}
+				if r == ' ' {
+					inDomain = false
+					sb.WriteRune(',')
+					sb.WriteRune(r)
+					continue
+				}
+			}
+		}
+		sb.WriteRune(r)
+	}
+	return sb.String()
 }
